+++ CHANGELOG.md
@@ -20,9 +20,23 @@
   - `addContactToChat()`
   - `deleteMessages()`
   - `getMessageInfo()`
+  - `getBasicChatInfo()`
+  - `marknoticedChat()`
+  - `getFirstUnreadMessageOfChat()`
+  - `markseenMsgs()`
+  - `forwardMessages()`
+  - `removeDraft()`
+  - `getDraft()`
+  - `miscSendMsg()`
+  - `miscSetDraft()`
 - jsonrpc: add `is_broadcast` property to `ChatListItemFetchResult` #3584
 - jsonrpc: add `was_seen_recently` property to `ChatListItemFetchResult`, `FullChat` and `Contact` #3584
-- jsonrpc: add `webxdc_info` property to `Message` #3588
+- jsonrpc: add `webxdc_info`, `parent_id` and `download_state` property to `Message` #3588
+- jsonrpc: add `BasicChat` object as a leaner alternative to `FullChat`
+- jsonrpc: add `last_seen` property to `Contact`
+- breaking! jsonrpc: replace `Message.quoted_text` and `Message.quoted_message_id` with `Message.quote`
+
+/todo add the rest
 
 ### Changes
 - order contact lists by "last seen";